events: Move common relation structs under events::room::encrypted to events::relation
