+++ crates/ruma-common/src/events/room/encrypted.rs
@@ -9,7 +9,12 @@ use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
 
 use super::message;
-use crate::{events::relation::InReplyTo, OwnedDeviceId, OwnedEventId};
+#[cfg(feature = "unstable-msc2677")]
+use crate::events::relation::Annotation;
+use crate::{
+    events::relation::{InReplyTo, Reference, Thread},
+    OwnedDeviceId, OwnedEventId,
+};
 
 mod relation_serde;
 
@@ -146,77 +151,6 @@ impl Replacement {
     /// Creates a new `Replacement` with the given event ID.
     pub fn new(event_id: OwnedEventId) -> Self {
         Self { event_id }
-    }
-}
-
-/// A reference to another event.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-pub struct Reference {
-    /// The event we are referencing.
-    pub event_id: OwnedEventId,
-}
-
-impl Reference {
-    /// Creates a new `Reference` with the given event ID.
-    pub fn new(event_id: OwnedEventId) -> Self {
-        Self { event_id }
-    }
-}
-
-/// An annotation for an event.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[cfg(feature = "unstable-msc2677")]
-#[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-pub struct Annotation {
-    /// The event that is being annotated.
-    pub event_id: OwnedEventId,
-
-    /// The annotation.
-    pub key: String,
-}
-
-#[cfg(feature = "unstable-msc2677")]
-impl Annotation {
-    /// Creates a new `Annotation` with the given event ID and key.
-    pub fn new(event_id: OwnedEventId, key: String) -> Self {
-        Self { event_id, key }
-    }
-}
-
-/// A thread relation for an event.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-pub struct Thread {
-    /// The ID of the root message in the thread.
-    pub event_id: OwnedEventId,
-
-    /// A reply relation.
-    ///
-    /// If this event is a reply and belongs to a thread, this points to the message that is being
-    /// replied to, and `is_falling_back` must be set to `false`.
-    ///
-    /// If this event is not a reply, this is used as a fallback mechanism for clients that do not
-    /// support threads. This should point to the latest message-like event in the thread and
-    /// `is_falling_back` must be set to `true`.
-    pub in_reply_to: InReplyTo,
-
-    /// Whether the `m.in_reply_to` field is a fallback for older clients or a real reply in a
-    /// thread.
-    pub is_falling_back: bool,
-}
-
-impl Thread {
-    /// Convenience method to create a regular `Thread` with the given event ID and latest
-    /// message-like event ID.
-    pub fn plain(event_id: OwnedEventId, latest_event_id: OwnedEventId) -> Self {
-        Self { event_id, in_reply_to: InReplyTo::new(latest_event_id), is_falling_back: false }
-    }
-
-    /// Convenience method to create a reply `Thread` with the given event ID and replied-to event
-    /// ID.
-    pub fn reply(event_id: OwnedEventId, reply_to_event_id: OwnedEventId) -> Self {
-        Self { event_id, in_reply_to: InReplyTo::new(reply_to_event_id), is_falling_back: true }
     }
 }
